\name{fg.adjust.inflation}
\alias{fg.adjust.inflation}
\title{
Adjusting the p-values based on the LR distribution
}
\description{
Adjusting the p-values based on the LR distribution. Notice:
it is applicable for the results of whole genome when the genome inflaion is not normal.
}
\usage{
fg.adjust.inflation(object)
}
\arguments{
  \item{object}{A S3 object of scaning result returned by \code{\link{fg.snpscan}}. }
}
\details{
}
\value{
A new S3 object with adjusted p-values.
}
\examples{
}
\keyword{ Scanning }
