\name{fg.get.pca}
\alias{fg.get.pca}
\title{
Calculating PCA matrix using PLINK software
}
\description{
Calculating PCA matrix for a genotype object using PLINK software.
}
\usage{
fg.get.pca(object, plink.path)
}
\arguments{
  \item{object}{S3 object \code{fgwas.gen.obj}, indicating the genotype object generated by the \code{\link{fg.load.plink}}, \code{\link{fg.load.simple}} or \code{\link{fg.simulate}}.}
  \item{plink.path}{string, indicating PLINK command path, for the large PLINK data, the package loads partial PLINK data extracted by the 'plink' command rather than the whole data set.}
}
\details{
This function returns a matrix containing 20 PCAs which can be used as the covariates in fGWAS. Please specify the location of PLINK in the function calling.
}
\value{
A matrix with 20 PCAs in columns. 
}
\examples{

obj1 <- fg.simulate( "Logistic", "AR1", 2000, 1000, 1:8, 
    phe.missing=0.05, snp.missing=0.05, 
    sig.pos=501, plink.format=FALSE, file.prefix = NULL );

m <- fg.get.pca( obj1$obj.gen, "~/lib/plink1.9/plink");
head(m)

#           PCA1         PCA2        PCA3       PCA4         PCA5         PCA6
#N_1  0.01379520 -0.017665100 -0.00619801 -0.0262379  0.008060390 -0.007591340
#N_2 -0.02162890 -0.000899263  0.01792060  0.0492322 -0.006867650 -0.007675600
#...

}
\keyword{ PCA }
