\name{fg.simulate}
\alias{fg.simulate}
\title{
Data Simulation
}
\description{
Data simulation is for demostration or power test of fGWAS package.
}
\usage{
fg.simulate(curve.type, covariance.type, n.obs, n.snp, time.points, 
    par0 = NULL, 
    par1 = NULL, 
    par2 = NULL, 
    par.covar = NULL, 
    par.X = NULL,
    sig.pos = NULL, 
    phe.missing = 0.03, 
    snp.missing = 0.03, 
    plink.format = FALSE, 
    file.prefix = NULL )
}
\arguments{
  \item{curve.type}{String, indicating the curve type, such as "Logistic", "Legendre2", available options are described in the details of \code{\link{fg.load.phenotype}} }
  \item{covariance.type}{String, indicating the covariance structure type, such as "SAD1", "AR1", available options are described in the details of \code{\link{fg.load.phenotype}}, . }
  \item{n.obs}{Numeric, indicating the individual number.}
  \item{n.snp}{Numeric, indicating the SNP number.}
  \item{time.points}{Numeric, indicating the measure time points for the longitudinal phenotype.}
  \item{par0}{Numeric vector, indicating the curve parameters for the genotype AA.}
  \item{par1}{Numeric vector, indicating the curve parameters for the genotype Aa.}
  \item{par2}{Numeric vector, indicating the curve parameters for the genotype aa.}
  \item{par.covar}{Numeric vector, indicating the parameters of the covariance matrix.}
  \item{par.X}{Numeric vector, indicating the covariate coefficients.}
  \item{sig.pos}{Numeric, indicating the significant SNP position.}
  \item{phe.missing}{Numeric, the missing rates of phenotype data.}
  \item{snp.missing}{Numeric, the missing rates of genotype data.}
  \item{plink.format}{Logical variable, indicating whether the PLINK data files are generated.}
  \item{file.prefix}{String, the prefix file name for the simulation data.}
}
\value{

The function return a list containing two objects, including\cr

  \item{obj.gen}{genotype data object, S3 object with class name 'fgas.gen.obj', the format is identical with return of \code{\link{fg.load.plink}} or \code{\link{fg.load.simple}}. }
  \item{obj.phe}{phenotype data object, S3 object with class name 'fgas.phe.obj', the format is identical with return of \code{\link{fg.load.phenotype}}. }

If \code{plink.format} is set to TRUE, the plink data set generated by the simulation precoess will be stored in the file specified by \code{obj.gen$files}.

}
\details{

The object structures of two S3 object are listed in \code{\link{fg.load.phenotype}} and \code{\link{fg.load.plink}} \cr

The mathematics model can be illustrated as the following formula: \cr

Y = X %*% \code{par.X} + curve_function ( \code{par0}, T ) + e if snp is AA \cr
Y = X %*% \code{par.X} + curve_function ( \code{par1}, T ) + e if snp is Aa \cr
Y = X %*% \code{par.X} + curve_function ( \code{par2}, T ) + e if snp is aa \cr

In above formula, X and snp are geneated randomly, curve_function is selected by the \code{curve.type}, the residuals \code{e} follow the multivarite normal distribution N(0, \code{covar} ), covar is built by the parameter \code{covariance.type} and \code{par.covar}.
}

\examples{

obj1 <- fg.simulate( "Logistic", "AR1", 2000, 1000, 1:8, phe.missing=0.05, snp.missing=0.05, sig.pos=501, plink.format=FALSE, file.prefix = NULL );
print(obj1$obj.gen);
print(obj1$obj.phe);
plot(obj1$obj.phe, curve.fitting=F);

obj2 <- fg.simulate( "Logistic", "SAD1", 2000, 800, 1:6, phe.missing=0.05, snp.missing=0.05, sig.pos=301, plink.format=TRUE, file.prefix = "test.fgwas" );
print(obj2$obj.gen);
print(obj2$obj.phe);
plot(obj2$obj.phe, curve.fitting=FALSE);

#remove the phenotype files and PLINK files
unlink(c("test.fgwas.geno.bed", "test.fgwas.geno.bed", "test.fgwas.geno.bed", "test.fgwas.pheT.csv", "test.fgwas.pheY.csv"));

}
\keyword{ Simulation }
