\name{fg.simulate}
\alias{fg.simulate}
\title{
Data Simulation
}
\description{
Data simulation is for demonstration or power test in the fGWAS package.
}
\usage{
fg.simulate(curve.type, covariance.type, n.obs, n.snp, time.points, 
    par0 = NULL, 
    par1 = NULL, 
    par2 = NULL, 
    par.covar = NULL, 
    par.X = NULL,
    sig.pos = NULL, 
    phe.missing = 0.03, 
    snp.missing = 0.03, 
    plink.format = FALSE, 
    file.prefix = NULL )
}
\arguments{
  \item{curve.type}{String, indicating the curve type, such as "Logistic", "Legendre2", available options are described in the details of \code{\link{fg.load.phenotype}}. }
  \item{covariance.type}{String, indicating the covariance structure type, such as "SAD1", "AR1", available options are described in the details of \code{\link{fg.load.phenotype}}. }
  \item{n.obs}{Numeric, indicating the individual number.}
  \item{n.snp}{Numeric, indicating the SNP number.}
  \item{time.points}{Numeric, indicating the measure time points for the longitudinal traits.}
  \item{par0}{Numeric vector, indicating the curve parameters for the genotype AA.}
  \item{par1}{Numeric vector, indicating the curve parameters for the genotype Aa.}
  \item{par2}{Numeric vector, indicating the curve parameters for the genotype aa.}
  \item{par.covar}{Numeric vector, indicating the parameters of the covariance matrix.}
  \item{par.X}{Numeric vector, indicating the covariate coefficients.}
  \item{sig.pos}{Numeric, indicating the significant SNP position.}
  \item{phe.missing}{Numeric, the missing rate of phenotype data.}
  \item{snp.missing}{Numeric, the missing rate of genotype data.}
  \item{plink.format}{Logical variable, indicating whether the PLINK data files are generated.}
  \item{file.prefix}{String, the prefix file name for the simulation data.}
}
\value{

The function return a list containing two objects, including\cr

  \item{obj.gen}{genotype data object, S3 object with class name \code{'fgas.gen.obj'}, the format is identical with return of \code{\link{fg.load.simple}}. }
  \item{obj.phe}{phenotype data object, S3 object with class name \code{'fgas.phe.obj'}, the format is identical with return of \code{\link{fg.load.phenotype}}. }

If \code{plink.format} is set to TRUE, the plink data set generated by the simulation process will be stored in the file specified by \code{obj.gen$files}.

}
\details{

The object structures of two S3 object, phenotype data object(\code{\link{fg.load.phenotype}}) and genotype object(\code{\link{fg.load.plink}}) \cr

The mathematics model can be illustrated as the following formulas: \cr

\preformatted{
Y = X \%*\% par.X + curve_function ( par0, T ) + e if snp is AA \cr
Y = X \%*\% par.X + curve_function ( par1, T ) + e if snp is Aa \cr
Y = X \%*\% par.X + curve_function ( par2, T ) + e if snp is aa \cr
}

In above formulas, X and snp are geneated randomly, curve_function is selected by the \code{curve.type}, the residuals \code{e} follow the multivarite normal distribution N(0, \code{covar} ), covariance matrix is built by the parameter \code{covariance.type} and \code{par.covar}.
}

\examples{

# simulate the Logistic traits and genotype data.
obj1 <- fg.simulate( "Logistic", "AR1", 2000, 1000, 1:8, 
    phe.missing=0.05, snp.missing=0.05, 
    sig.pos=501, plink.format=FALSE, file.prefix = NULL );

# show the genotype object    
print(obj1$obj.gen);
# show the phenotype object    
print(obj1$obj.phe);
# plot the phenotype object    
plot(obj1$obj.phe, curve.fitting=FALSE);

# simulate the Logistic traits and genotype data.and save the data into PLINK files
obj2 <- fg.simulate( "Logistic", "SAD1", 2000, 800, 1:6, 
    phe.missing=0.05, snp.missing=0.05, sig.pos=301, 
    plink.format=TRUE, file.prefix = "temp.fgwas" );
    
    
# show the genotype object    
print(obj2$obj.gen);
# show the phenotype object    
print(obj2$obj.phe);
# plot the phenotype object    
plot(obj2$obj.phe, curve.fitting=FALSE);

# remove the phenotype files and PLINK files
unlink(c("temp.fgwas.geno.bed", "temp.fgwas.geno.bed", "temp.fgwas.geno.bed", 
    "temp.fgwas.pheT.csv", "temp.fgwas.pheY.csv"));

}
\keyword{ Simulation }
