\name{fg.simulate}
\alias{fg.simulate}
\title{
Data Simulation
}
\description{
Data simulation is for the test the power of fGWAS package.
}
\usage{
fg.simulate(curve.type, covariance.type, n.obs, n.snp, time.points, 
    par0 = NULL, 
    par1 = NULL, 
    par2 = NULL, 
    par.covar = NULL, 
    par.X = NULL,
    sig.pos = NULL, 
    phe.missing = 0.03, 
    snp.missing = 0.03, 
    plink.format = FALSE, 
    file.prefix = NULL )
}
\arguments{
  \item{curve.type}{String, indicating the curve type, available options are "Logistic", "DoubleLogistic", "ABRK". }
  \item{covariance.type}{String, indicating the covariance structure type, available options are "SAD1", "AR1". }
  \item{n.obs}{Numeric, indicating the individual number.}
  \item{n.snp}{Numeric, indicating the SNP number.}
  \item{time.points}{Numeric, indicating the measure time points for the longitudinal phenotype.}
  \item{par0}{Numeric vector, indicating the parameters of the curve for the genotype AA.}
  \item{par1}{Numeric vector, indicating the parameters of the curve for the genotype Aa.}
  \item{par2}{Numeric vector, indicating the parameters of the curve for the genotype aa.}
  \item{par.covar}{Numeric vector, indicating the parameters of the covariance matrix.}
  \item{par.X}{Numeric vector, indicating the covariate coefficients.}
  \item{sig.pos}{Numeric, indicating the significant SNP position.}
  \item{phe.missing}{Numeric, the missing rates of phenotype data.}
  \item{snp.missing}{Numeric, the missing rates of genotype data.}
  \item{plink.format}{Logical variable, indicating whether the PLINK data files are generated.}
  \item{file.prefix}{String, the prefix file name for the simulation data.}
}
\value{
A list with two objects is returned by this function, including\cr

  \item{obj.gen}{S3 object with class name 'fgas.gen.obj', indicating the genotype data.}
  \item{obj.phe}{S3 object with class name 'fgas.phe.obj', indicating the phenotype data.}

if \code{plink.format} is set to TRUE, the plink file names generated by the simulation precoess will be stored in the \code{obj.phe$files}.

}
\details{
Y = X + curve;
}
\references{
}
\seealso{
See the \code{\link{fg.load.phenotype}} to check the object structure of 'fgas.phe.obj'.

}
\examples{

}
\keyword{ fGWAS }
\keyword{ Simulation }
