\name{fg.snpscan}
\alias{fg.snpscan}
\title{
Scaning SNP
}
\description{
Scaning SNP with the fGWAS model and detect the significant SNPs.
}
\usage{
fg.snpscan( fgwas.gen.obj, fgwas.phe.obj, 
    method = "optim-fgwas", 
    curve.type = NULL, 
    covariance.type = NULL, 
    snp.sub = NULL, 
    options = list() )
}
\arguments{
  \item{fgwas.gen.obj}{S3 object \code{fgwas.gen.obj}, indicating the phenotype object generated by the \code{\link{fg.load.phenotype}} or \code{\link{fg.simulation}}.}
  \item{fgwas.phe.obj}{S3 object \code{fgwas.phe.obj}, indicating the genotype object generated by the \code{\link{fg.load.plink}}, \code{\link{fg.load.simple}} or \code{\link{fg.simulation}}.}
  \item{method}{string, indicating the model used in the scaning process, available options are "gls", "mixed", "fast", "fgwas", "optim-fgwas". Default is 'optim-fgwas'.}
  \item{curve.type}{string, indicating the curve type used in the scaning process for the method "fast", "fgwas" and "optim-fast", instead of the curve type in data object. The full optional values are listed in \code{\link{fg.load.phenotype}}}
  \item{covariance.type}{string, indicating the type of covariance structure used in the scaning process for the method "fast", "fgwas" and "optim-fast", instead of the curve type in data object. The full optional values are listed in \code{\link{fg.load.phenotype}}}
  \item{snp.sub}{vector of SNP index or SNP name, indicating the selected SNPs will be scanned, not the whole data set.}
  \item{options}{list}
}
\value{

A S3 object (\code{fgwas.scan.obj}) is returned by this function, including:\cr

  \item{obj.gen}{S3 object \code{fgwas.gen.obj} processed in this function}
  \item{obj.phe}{S3 object \code{fgwas.phe.obj} processed in this function}
  \item{ret.gls}{Matrix, results for each SNP scanned by the GLS method.}
  \item{ret.mixed}{}
  \item{ret.fast}{Matrix, results for each SNP scanned by the FAST method.}
  \item{ret.fgwas}{Matrix, results for each SNP scanned by the fGWAS method.}


\code{ret.gls} includes the following information: \cr

  \item{INDEX  }{SNP Index}
  \item{NAME   }{SNP Name}
  \item{CHR    }{Chromosome}
  \item{POS    }{Base position}
  \item{NMISS  }{Number of missing.}
  \item{MAF    }{Minor Allel frequence}
  \item{LR     }{p-value of Fisher .}
  \item{pv     }{p-value of Fisher .}


\code{ret.mixed} includes the following information: \cr

  \item{INDEX  }{SNP Index}
  \item{NAME   }{SNP Name}
  \item{CHR    }{Chromosome}
  \item{POS    }{Base position}
  \item{Allel1 }{Allel 1}
  \item{Allel2 }{Allel 2}
  \item{MAF    }{Minor Allel frequence}
  \item{NMISS  }{Number of missing.}
  \item{pv     }{p-value of Fisher .}
  \item{p.min  }{p-value of minimum values.}
  \item{p.join }{p-value of joint value.}
  \item{p0     }{p-value of 0th-degree legendre polynomial.}
  \item{p1     }{p-value of 1st-degree legendre polynomial.}
  \item{p2     }{p-value of 2nd-degree legendre polynomial.}
  \item{p3     }{p-value of 3rd-degree legendre polynomial.}
  \item{p4     }{p-value of 4th-degree legendre polynomial.}

  
\code{ret.fgwas} and \code{ret.fast}  includes the following information:\cr

  \item{INDEX  }{SNP Index}
  \item{NAME   }{SNP Name}
  \item{CHR    }{Chromosome}
  \item{POS    }{Base position}
  \item{MAF    }{Minor Allel frequence}
  \item{NMISS  }{Number of missing.}
  \item{SNP0   }{Individual Count of genotype 0}
  \item{SNP1   }{Individual Count of genotype 1}
  \item{SNP2   }{Individual Count of genotype 2}
  \item{GENO   }{Genotype count}
  \item{LR     }{Likelihood Ratio.}
  \item{pv     }{p-value based on LR.}
  \item{h0.X0  }{the 1st covariate coefficient under NULL hypothesis}
  \item{...    }{the other covariate coefficients under NULL hypothesis}
  \item{h0.a   }{the parameter 'a' in curve function( e.g. logistic curve) under NULL hypothesis}
  \item{h0.b   }{the parameter 'b' in curve function( e.g. logistic curve) under NULL hypothesis}
  \item{...    }{the other curve parameters in the curve function under NULL hypothesis}
  \item{h0X.rho}{the parameter 'rho' in covariance structure ( e.g. AR1 ) under NULL hypothesis}
  \item{...    }{the other parameters in covariance structure ( e.g. AR1 ) under NULL hypothesis}
  \item{h1.X0  }{the 1st covariate coefficient under alternative  hypothesis}
  \item{...    }{the other covariate coefficients under alternative  hypothesis}
  \item{h1.G0.a}{the parameter 'a' for genotype 0 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1.G0.b}{the parameter 'b' for genotype 0 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{...    }{the other parameters for genotype 0 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1.G1.a}{the parameter 'a' for genotype 1 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1.G1.b}{the parameter 'b' for genotype 1 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{...    }{the other parameters for genotype 1 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1.G2.a}{the parameter 'a' for genotype 2 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1.G2.b}{the parameter 'a' for genotype 2 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{...    }{the other parameters for genotype 2 in curve function (e.g. logistic curve) under alternative hypothesis}
  \item{h1X.rho}{the parameter 'rho' in covariance structure ( e.g. AR1 ) under alternative hypothesis}
  \item{...    }{the other parameters in covariance structure ( e.g. AR1 ) under alternative hypothesis}


}

\details{
If the curve type 
}

\seealso{
\code{\link{fg.load.phenotype}}, \code{\link{fg.load.plink}}, \code{\link{fg.load.simple}}, \code{\link{fg.select.sigsnp}}
}
\examples{
}
\keyword{ fGWAS }
\keyword{ Scaning }
