\name{fg.snpscan}
\alias{fg.snpscan}
\title{
Detecting significant SNPs by fGWAS model
}
\description{
Scaning SNPs with the fGWAS model and detect the significant SNPs.
}
\usage{
fg.snpscan( fgwas.gen.obj, fgwas.phe.obj, 
    method = "optim-fgwas", 
    curve.type = NULL, 
    covariance.type = NULL, 
    snp.sub = NULL, 
    options = list(verbose=F) )
}
\arguments{
  \item{fgwas.gen.obj}{S3 object \code{fgwas.gen.obj}, indicating the genotype object generated by the \code{\link{fg.load.plink}}, \code{\link{fg.load.simple}} or \code{\link{fg.simulate}}.}
  \item{fgwas.phe.obj}{S3 object \code{fgwas.phe.obj}, indicating the phenotype object generated by the \code{\link{fg.load.phenotype}} or \code{\link{fg.simulate}}.}
  \item{method}{string, indicating the statistical model used in the scaning process, available options are "gls", "mixed", "fast", "fgwas", "optim-fgwas". Default is 'optim-fgwas'.}
  \item{curve.type}{string, indicating the curve type used in the scaning process for the method "fast", "fgwas" and "optim-fast", instead of the original curve type in phenotype object. The full optional values are listed in \code{\link{fg.load.phenotype}}}
  \item{covariance.type}{string, indicating the type of covariance structure used in the scaning process for the method "fast", "fgwas" and "optim-fast", instead of the curve type in phenotype object. The full optional values are listed in \code{\link{fg.load.phenotype}}}
  \item{snp.sub}{vector of SNP index or SNP name, indicating the selected SNPs will be scanned, not the whole data set.}
  \item{options}{list}
}
\value{

A S3 object with the class name (\code{fgwas.scan.obj}) is returned by this function, including:\cr

  \item{obj.gen}{S3 object \code{fgwas.gen.obj} processed in this function}
  \item{obj.phe}{S3 object \code{fgwas.phe.obj} processed in this function}
  \item{ret.gls}{Matrix, results for each SNP scanned by the 'gls' method.}
  \item{ret.mixed}{Matrix, results for each SNP scanned by the 'mixed' method.}
  \item{ret.fast}{Matrix, results for each SNP scanned by the 'fast' method.}
  \item{ret.fgwas}{Matrix, results for each SNP scanned by the 'fgwas' or 'fgwas-optim' method.}

If 'gls' is used to scan the SNPs, the result is stored in \code{ret.gls} with the following items:\cr

  \item{INDEX  }{SNP index.}
  \item{NAME   }{SNP name.}
  \item{CHR    }{Chromosome.}
  \item{POS    }{Base position.}
  \item{NMISS  }{Number of missing.}
  \item{MAF    }{Minor allel frequence.}
  \item{LR     }{Likelihood ratio.}
  \item{pv     }{p-value of Fisher.}


If 'mixed' is used to scan the SNPs, the result is stored in \code{ret.mixed} with the following items:\cr

  \item{INDEX  }{SNP index.}
  \item{NAME   }{SNP name.}
  \item{CHR    }{Chromosome.}
  \item{POS    }{Base position.}
  \item{Allel1 }{Allel 1.}
  \item{Allel2 }{Allel 2.}
  \item{MAF    }{Minor allel frequence.}
  \item{NMISS  }{Number of missing.}
  \item{pv     }{p-value of Fisher .}
  \item{p.min  }{p-value of minimum values.}
  \item{p.join }{p-value of joint value.}
  \item{p0     }{p-value of 0th-degree Legendre polynomial.}
  \item{p1     }{p-value of 1st-degree Legendre polynomial.}
  \item{p2     }{p-value of 2nd-degree Legendre polynomial.}
  \item{p3     }{p-value of 3rd-degree Legendre polynomial.}
  \item{p4     }{p-value of 4th-degree Legendre polynomial.}

  
  If 'fgwas', 'fgwas-optim' or 'fast' are  used to scan the SNPs, the result is stored in \code{ret.fast} or \code{ret.fgwas} with the following items:\cr

  \item{INDEX  }{SNP index.}
  \item{NAME   }{SNP name.}
  \item{CHR    }{Chromosome.}
  \item{POS    }{Base position.}
  \item{MAF    }{Minor allel frequence.}
  \item{NMISS  }{Number of missing.}
  \item{SNP0   }{Individual count of genotype 0.}
  \item{SNP1   }{Individual count of genotype 1.}
  \item{SNP2   }{Individual count of genotype 2.}
  \item{GENO   }{Genotype count.}
  \item{LR     }{Likelihood Ratio.}
  \item{pv     }{p-value based on LR.}
  \item{h0.X0  }{the 1st covariate coefficient under NULL hypothesis.}
  \item{...    }{the other covariate coefficients under NULL hypothesis.}
  \item{h0.a   }{the parameter 'a' in curve function( e.g. Logistic curve) under NULL hypothesis.}
  \item{h0.b   }{the parameter 'b' in curve function( e.g. Logistic curve) under NULL hypothesis.}
  \item{...    }{the other curve parameters in the curve function under NULL hypothesis.}
  \item{h0X.rho}{the parameter 'rho' in covariance structure ( e.g. AR1 ) under NULL hypothesis.}
  \item{...    }{the other parameters in covariance structure ( e.g. AR1 ) under NULL hypothesis.}
  \item{h1.X0  }{the 1st covariate coefficient under alternative  hypothesis.}
  \item{...    }{the other covariate coefficients under alternative  hypothesis.}
  \item{h1.G0.a}{the parameter 'a' for genotype 0 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{h1.G0.b}{the parameter 'b' for genotype 0 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{...    }{the other parameters for genotype 0 in curve function (e.g. logistic curve) under alternative hypothesis.}
  \item{h1.G1.a}{the parameter 'a' for genotype 1 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{h1.G1.b}{the parameter 'b' for genotype 1 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{...    }{the other parameters for genotype 1 in curve function (e.g. logistic curve) under alternative hypothesis.}
  \item{h1.G2.a}{the parameter 'a' for genotype 2 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{h1.G2.b}{the parameter 'b' for genotype 2 in curve function (e.g. Logistic curve) under alternative hypothesis.}
  \item{...    }{the other parameters for genotype 2 in curve function (e.g. logistic curve) under alternative hypothesis.}
  \item{h1X.rho}{the parameter 'rho' in covariance structure ( e.g. AR1 ) under alternative hypothesis.}
  \item{...    }{the other parameters in covariance structure ( e.g. AR1 ) under alternative hypothesis.}
  \item{h0.R2  }{R2 under NULL hypothesis.}
  \item{h1.R2  }{R2 under alternative hypothesis.}

}

\details{

}

\seealso{
\code{\link{fg.load.phenotype}}, \code{\link{fg.load.plink}}, \code{\link{fg.load.simple}}, \code{\link{fg.select.sigsnp}}
}
\examples{

# simulate the phenotype object and genotype object
r<-fg.simulate("Logistic", "AR1", 2000, 500, 1:7, sig.pos=250 );
# SNP scaning in a short range (245:255) using 'fast' method
obj1.scan <- fg.snpscan(r$obj.gen, r$obj.phe, method="fast", snp.sub=c(245:255) );
# show the summary information of result object
obj1.scan;

# SNP scaning in a short range (245:255) using 'fgwas' method
obj2.scan <- fg.snpscan(r$obj.gen, r$obj.phe, method="fgwas", snp.sub=c(245:255) );
obj2.scan;

# check the full result table.
tb.full <- obj2.scan$ob.fgwas$results;

# plot Manhattan figure
plot(obj2.scan, file.pdf="temp.fwgas.obj2.scan.pdf");

# select significant SNPs
tb.sig <- fg.select.sigsnp(obj2.scan, sig.level=0.001,  pv.adjust = "bonferroni")
# plot the genetic curve effect for the significant SNPs
plot.fgwas.curve( obj2.scan, tb.sig$INDEX, file.pdf="temp.fwgas.obj2.curve.pdf");

}
\keyword{ Scanning }
