\name{plot.fgwas.curve}
\alias{plot.fgwas.curve}
\title{
Plot the curve associated with genotype
}
\description{
Plot the curve associated with genotype
}
\usage{
plot.fgwas.curve(object, snp.sub, file.pdf = NULL)
}
\arguments{
  \item{object}{Object returned by the \code{\link{fg.snpscan}}}
  \item{snp.sub}{Vector of SNP index or SNP name.}
  \item{file.pdf}{string indicating the PDF file name.}
}
\details{
}
\value{
No return values.
}
\examples{
}
\keyword{ plot }
